moved contents of index.js to correct location
